requirements.txt
discord.py==1.6.0
pytz==2021.1